Cleaned: created two new header file to split the function and struc declaration
