+++ includes/minishell.h
@@ -35,108 +35,9 @@
 /*Include libft*/
 # include "../libft/libft.h"
 
-typedef enum e_error
-{
-	SUCCESS = 0,
-	ERR_LAUNCH = 1,
-	ERR_MALLOC = 2,
-	ERR_PARCING = 3,
-	ERR_ENV = 4,
-	//..add any other relevant error
-	//do not forget to add them in print_error function in errors.c
-}	t_error;
-
-//When you add token type, add it in the utils to print file as well
-typedef enum e_token_type
-{
-	TYPE_ARG, //all regular words or arguments
-	TYPE_COMMAND, //built-in commands
-	TYPE_DQUOTE, // double quote
-	TYPE_SQUOTE, //single quote
-	TYPE_PIPE, // pipe
-	TYPE_REDIRIN, //< (input redirection)
-	TYPE_REDIROUT, // > (output redirection)
-	TYPE_REDIRAPPEND, // >> (append output redirection)
-	TYPE_HEREDOC_OP, // << (here document)
-	TYPE_HEREDOC_DELIM, // heredoc delimiter
-	TYPE_HEREDOC_CONTENT, // heredoc content in between delimiters
-	TOKEN_EMPTY,
-}			t_token_type;
-
-typedef struct s_env
-{
-	char			*key;
-	char			*value;
-	struct s_env	*next;
-}			t_env;
-
-typedef struct s_token
-{
-	t_token_type	type;
-	char			*content;
-	struct s_token	*next;
-}			t_token;
-
-//Used during token extraction to store
-//quote information and pass it to different functions
-typedef struct s_quote_info
-{
-	int		in_inner_quote;
-	char	inner_quote;
-	char	outer_quote;
-	int		num_outer_quote;
-}			t_quote_info;
-
-//Used during token expansion to pass more than 4 args
-typedef struct s_parse_params {
-	size_t	i;
-	size_t	j;
-	int		in_squote;
-	int		in_dquote;
-}	t_parse_params;
-
-//Main Shell structure, to enhanced and figure out how to use it
-typedef struct s_shell
-{
-	t_env	*env_list;
-	t_token	*token_list;
-	int		exit_status;
-}	t_shell;
-
-//PARCING
-//token_creation.c
-t_error				input_to_token(t_token **token_list, char *args);
-t_token				*create_token(char *input, t_token_type type);
-t_error				add_token(t_token **token_list, \
-					char *input, t_token_type type);
-t_token_type		get_token_type(char *input, t_token **token_list);
-int					is_command(t_token **token_list);
-
-//token_extraction.c
-t_error				ft_split_token(t_token **token_list, \
-						char *args, size_t *i, char **token);
-
-//token_expansion.c
-t_error				expand_tokens(t_token **token_list, t_env *env_list);
-
-//token_extraction_utils.c
-int					is_space(char c);
-int					is_delimiter(char c);
-int					get_delimiter_len(const char *str);
-
-//token_expansion_var.c
-char				*get_var_value(t_env *env_list, char *key);
-char				*get_var_name(const char *str);
-void				get_var_length(const char *str, \
-						t_env *env_list, size_t *i, size_t *expanded_len);
-void				expand_var(const char *str, \
-						t_env *env_list, size_t *i, char *expanded);
-
-//token_expansion_quotes.c
-void				init_expansion_params(size_t *i, size_t *j, \
-						int *in_squote, int *in_dquote);
-void				process_quote(char quote_char, int *in_squote, \
-						int *in_dquote);
+/*Include other headers*/
+# include "parcing.h"
+# include "types.h"
 
 //ENV
 //init_env.c
